Add Concourse task for CI
To add this repository to our CI pipelines we need to define a Concourse task
in ci/integration.yml. For now we don't run the tests on our CI server as they
require a Postgres instance running and we would need to build a custom Docker
image for this. Currently we'll depend on Travis running the tests for us.