RepairEvents function
We are missing events for services or applications that have never been
started/updated/stopped since we started collecting data.

RepairEvents will emit START events for any apps or services that cloud
controller reports as running but that we have no record of.

It is comparable to the cf api's
"destructively_purge_all_and_reseed_started_apps" usage events api, but
it is non-dustructive and reversable (since all 'fake' events are
created with id=0 so they can be removed.