Add test helper for create/destroy test databases
We want to ensure we are running against a clean db during testing. To
do this we create a database and return a connection string that can be
used to run migrations against, then tear it all down cleanly.

This is faster than tearing down the entire postgres server and less
complicated than ensuring all the objects are removed between tests or
converting all code to work in a sinle tx.