Add db.QueryRowJSON and db.QueryJSON methods
QueryJSON will execute an SQL query against the database and stream the
results as JSON directly to a io.Writer.

Each row directly to the writer w as JSON without needing to Scan into
intermediate vars or do any marshalling. This allows very large queries
to be streamed from the database to the writer effciently.

An UpdateViews method was added to trigger updates of the materialized
views.