Use UAA for authorization
UAA tokens are now required for allow requests. You can authorize the
app via oauth by visiting /oauth/authorize

This ensures that usage data can only be fetched for spaces that the
user has access to.

Currently we use the "cloud_controller.admin" scope to indicate if the
user is allowed to update pricing_plans. In the future it may make sense
to create a new scope in UAA ie "billing.write" for this purpose.