Optimisation: Use UNION ALL and add 'state' index
UNION queries were mistakenly used to combine the app/service events.
UNION ALL should be preferred for this use case since we do not need to
de-dup the results.

Adding an index on the 'state' value in the raw_message json gives a
significant speed increase in recalculating the billable view.