Complete implementation of RepairEvents
The previous RepairEvents function was incomplete it ahs been extended
to cover the following cases:

* No recorded event for an app we know is running
* No recorded event for a service we know is running
* Missing start event for app we have seen a STOPPED event for
* Missing start event for service we jave seen a DELETED event for

The follow cases are intensionally ignored:

* If we saw a service UPDATED event but without a CREATED: This is
ignored as we cannot predict the previous state data
* If we saw an app update event (STARTED with previous state STARTED):
This is ignore as we do not consider the impact to be too servere.

A cli tool `db_repair` has been created to allow running the
RepairEvents function as a standalone tool.