/report: Filter by single org GUID
So that we can generate reports for individual orgs. This should speed up
the query a bit because we can filter earlier on. The HTML version of the
report can be exported as a PDF and sent to a tenant.

We're returning `org_guid` as a constant value in the SQL statement because
we can't figure out how to extract it from the query path in the template.