/report: Summarise by resource name
So that an app that is regularly blue-green deployed only has a single line
entry in the bill. We've added additional fixtures for the first app to test
that they are summed into a single entry.

The original name that an app or service was created with is used in the
report, as this should make most sense to tenants and is easiest for us to
generate. We've amended the names of the apps in the stop events to indicate
this behaviour.

The details links and column have been removed from the report because there
is no longer a single resource GUID that they are related to. For this
reason we no longer need to return the GUID.

Now that we're using the names from the usage events we no longer need to
query the CF API for app and service instance names when generating the
report.