/report: Use correct path in form submission
Previously this was submitting to the old endpoint name without the
originally requested org GUID.

We couldn't use `c.Path()` because that contains `/report/:org_guid` whereas
`c.Request().URL.Path` contains the actual request path.

Return the latter for `Title` and `Path` because it fixes the original issue
and is more consistent when we show this to people via the templates.