/report: Return message when org not found
Previously this would generate a server-side error and abruptly finish the
HTML output because the org total was JSON `null` and there were no
arguments when converting the price:

    <in_pounds>: wrong number of args for in_pounds: want 1 got 0

Fix this by generating a different message if there are no spaces. This
isn't a problem for spaces because the range is a noop if there are no
spaces.