Update reference formulas to AWS base costs
We are going to start by generating bills based on just the base AWS
costs so are updating the formulas to the hourly costs for the ireland
region.

Compute plan pricing is now calculated as:

    r4.xlarge (cell):
            cost per hour = 0.2960
            memory = 30.5

    0.2960 / 30.5 = 0.009704918032786884

                  = ~ 0.01 / GB / Hour

The RDS costs have been updated to match those listed on the AWS RDS
pricing page for each database type.