Add hacky script to push in batch pricing policies
Small hacky script to push in batch all the pricing policies
from a pricing.json file. Currently these policies can
be updated manually with the simple UI provided by the collector
but that is tedious and we do not track the policy changes.

As a temporary solution, we would push the formulas using a small
script. In the future this should be replaced by an endpoint that
allows to push all the policies in one go.

WARNING This script might break the collector, by creating a deadlock.
See  story https://www.pivotaltracker.com/story/show/154395415 for
more info. Solution: Wait a while, restart collector.