Script to generate monthly pdf bills in batch
In the sort term, we want to generate the bills in PDF with some
temporary solution based on the collector reports.

We want the bills from the orgs that are NOT in the "default" quota,
as we consider that every org that is not in trial would not use the
default quota. There are a pair of exceptions that we would handle
manually.

Here we write a hacky script that would directly download the report
in simple HTML for every org that is not in the default quota. Then
we convert the HTML to PDF using the tool wkhtmltopdf,