Do not UpdateView after updating pricing_plans
Executing the UpdateView call after every modification of the
pricing_plans via the API can cause problems when these calls are
performed in quick succession.  There is a known issue that can cause
the database to lock during view updates and this exhasperates the
problem.

We remove these calls, which means you must wait until the view is
refreshed (1hr) until the view is updated .