Add a hardcoded note about exchange rate to output
We are currently unsure of the direction we will take regarding billing
in USD or GBP.

For now we have hardcoded the formulas to include an exchange rate so
that the rendered outputs are in GBP.

We want to add a small note on the output so the exchange rate used is
mentioned.

This note should be removed once there is a decision about what
currency we are billing in or long-term solution for managing exchange
rates is in place.