Move pricing computation to billable_range func
We want to decouple the generation of the view "billable" with all the
events from the correlation and computation of the final pricing using
the formulas.

This way we do not need to recompute the view whenever we change the
formulas and plans. Additionally, it would later allow us to compute the
pricing using other tables or views, for instance for the pricing
estimation endpoints.

We join with the pricing plans and compute the formulas in the
billable_range function itself.

Note: The pricing in the billable view was actually not being used at
all.