Move billable_range function to inline query
To be able to do pricing simulations, we will want to use the logic in
the billable_range() SQL function with different tables. But we did not
find a sensible way to parametrise tables in SQL functions.

The code in billable_range() is actually a select query can we can move
it back to the API controller, and use the templating capabilities from
golang to build the query.