Use echo.Group for auth middlewares
We want to introduce a new endpoint that doesn't require authentication.
It will accept some fake events and return a price for them, so it
doesn't need to talk to Cloud Foundry and won't have access to any
sensitive data.

The existing `echo.Use` applied the authentication middleware to all
endpoints and doesn't allow use to have an endpoint that doesn't require
authentication.

So we've used `echo.AuthGroup` so setup a group for auth and a group for
admin auth. Endpoints added outside of these groups do not use auth. The
empty prefix is a bit of a hack - it seems that it's intended to be use
in a hierarchical way, but it performs simple string concatenation on
the paths, so we can still have all these endpoints hanging off the root
path.