Expose DB transaction methods on SQLClient
This allows an existing connection to be upgraded to use a transaction.
We need this so that we can create a temporary table within a report
handler, that contains fake events, use that table in the report query,
and then discard it afterwards.

The mocks were regenerated using the Make target described in the README.