Add simulated report endpoint
This will be used by the pricing calculator app to get cost estimations
against our real pricing plans for a set of apps and services over a
time range that the user has submitted.

We are using the existing (albeit refactored) queries in the database,
even though a simpler implementation could have been achieved without,
because we want to make sure that the pricing calculator and billing
reports use the same logic as they may get more complex in future (e.g.
calculating currency conversion or tax).

We are populating a temporary table with the submitted events and using
that in place of the `billable` materialized view. The table is created
within a transaction and the transaction is passed onto the report
function so that it has access to the table, then the transaction is
rolled back so that the table is discarded. It's for this reason that
we've had to paramaterise the view/table name and pass it around an
awful lot.

The endpoint is unauthenticated because it doesn't need access to Cloud
Foundry or any sensitive data.