Add the OAuth endpoints to the root middleware
We were catching the paths for oauth authentication
`/oauth/authorize` and `/oauth/callback` in the
UAATokenAuthentication middleware.

But now that middleware is defined using a group, so it
won't be used for those paths.

To fix that we will define the routes explicitly instead,
using the corresponding handlers from `authority`