Replace go-bindata
Load files straight from disk since we push the app with source code to
Cloud Foundry. Having separate `.sql` files is useful to take advantage
of tooling like syntax highlighting, linting and psql while developing

We default to finding files at a default location based on the `$PWD`
environment variable which is set by most shells as the working
directory and the go_buildpack/cloudfoundry sets it to the location used
at staging time.

We also expose a new configuration variable DATABASE_SCHEMA_DIR that can
be set to override the default in case you are running in an environment
that does not set `$PWD` (like travis) and so we can work around the go
test runner that likes to move the working dir.