Refactor collector
The collector contained too much logic and was running two event fetchers. This
made the tests incredibly complex and it's hard to introduce a third event
collector.

I'm introducing an EventFetcher interface which is solely responsible for
fetching the next batch of events and the collector manages the scheduling and
error handling.