Implement the Compose events collector
The Compose API provides the audit events through the /audit_events endpoint [1]. The API lists the events in chronological descending order and we can use the cursor parameter for proper pagination. This behaviour is different from the Cloud Foundry usage events APIs which lists the events in ascending order and makes the event fetching trivial.

In this case the fetching algoritm is the following:
 - we store the latest event id in the compose_audit_events_cursor table with
   the "latest_event_id" key
 - we store the cursor (last event's id on the current page) if we got a full
   page back from the API
 - if the stored cursor is nil we fetch the latest events and we update the
   latest_event_id in the database
 - if the stored cursor is not nil we fetch the events after that cursor (next
   page) and we don't touch the latest_event_id
 - when we get the next batch of data we check if we reached the
   latest_event_id (or there is no more events) and we stop there
 - we save only specific events, where event.Event is "deployment.scale.members"

[1] https://apidocs.compose.com/reference#2016-07-get-audit_events