Record Memory and Storage in pricing_plans
We add `memory_in_mb` and `storage_in_mb` columns to the `pricing_plans`
table because we are not currently storing this information, and it
is not provided in the service usage events. We chose adding columns
to this table because this data is naturally coupled to each plan (by
GUID), and if the plan changes a new row can be added.

Unfortunately the `IF NOT EXISTS` clause for adding a column was
introduced in Postgres 9.6, and we are running 9.5, so we have to catch
exceptions instead.  This should be refactored when we are using 9.6.

We also added a mechanism for testing migrations. In order to test a
migration we needed a way of only applying some of the migrations.  This
allows us to do things like run some migrations, load data, then apply
the migration and assert we didn't break anything.