Refactor `billable` materialized view
This will now need to be pull memory and storage data from columns
added to the `pricing_plans` table in a later migration. Because
these columns won't exist at this point we have to defer the creation
of this materialized view until later in the migration.

This also gives us an opportunity to rename it from `billable` to
something else, because it doesn't really represent a billable entity
anymore. It is now more like the duration a resource was used for.