Include memory and storage in pricing plans
Now these are stored in the `pricing_plans` table we can start using
them in the billing calculations.

We use `coalesce` in the `monetized_resources` common table expression
because apps and services provide this data in different ways. For apps
this data is already available, so we try to use that first. For
services this will be `NULL`, so we get it from the pricing plan.