Implement currency support per component
We want to allow the different components to have different
currencies. The base currency would be GBP, and we would apply
the exchange rate for each component when generating the
report and applying the formulas.

In this case we implement the `currency_rates` table in a similar
way than the `pricing_plans` with a `valid_from` date to set
from when the currency is valid.

In the SQL definition we add the new column with a default value
of `GBP`, then we drop the default definition. To simplify the
computation, `GBP` is a currency with default exchange rate of 1.

The currency code must be defined within a set of valid accepted
currencies (currently EUR, USD and GBP). Currency entries
can be repeated with different `valid_from` values.

In the monetise query we use the WINDOW operator to compute
the valid range for each exchange rate, defined by the valid_from
value and the next valid currency entry.

The pricing plan components have a reference to the currency
code to apply.

In the `monetized_resources` we would join the matching currency
entries which have a valid time frame that intersect with the queried
time window, the event duration and the pricing plan valid time window.

In the tests, we use the formula with a fixed rate added that
is apart of the time in seconds, to ensure that we test cases where
we incorrectly compute intersections with a range of 0 seconds.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>