Run currency conversion tests against the /events/raw endpoint
We want to validate that the usage events are properly generated, but we loose
information if we use any other endpoints which use some form of group by.

For example to validate the join conditions we need to see whether we generate
usage events with zero duration, but these events would be hidden because of
the group bys.