Fix missing named return value in migration code
The ApplyMigrations function that handles executing the scripts relies
on the behaviour of a named return value "err" being set to determin if
the transaction is committed or rolledback during a deferred function.

The function signature was missing the named err value and so even if an
error occured it would still attempt to commit the transaction, which
was not the intension.

Naming the return value resolves the problem.