Always use timestamptz not timestamp
I had an issue when running tests that turned out to have something to
do with how locales were setup in the VM I was running on, which caused
the casting of timestamp -> timestamptz to drift.

timestamp without time zone should not be used since how it is returned
and evaluated is complicated to reason about (based on the locale
settings of both the server and the client that is connected to it).

We originally thought we had worked around the issue by simply casting
to timestamptz in our VIEW of the data, but the complexities have still
managed to surface.

To avoid this in the columns have been altered to be timestamptz fields
cast with the AT TIME ZONE 'UTC' that makes sure they migrate correctly