Fix creation of new vat_rates on InitSchema
An issue was introduced that caused a new pair of vat_rates (Standard,
Zero) to be inserted into the database each time the InitSchema
"migration" ran.

The intension was to configure two default rates if
none were currently confgiured to bootstrap the process since it is
unlikely these will change for quite some time. However the method used
did not take into account the auto increment id so it just kept creating
more.

To resolve this we insert the rows with a known id and bump up the
auto-increament sequence to start handing out ids from 101, this gives
plenty of room for any other hard coded values we want to use in tests
etc and means we can use the ON CONFLICT (id) as intended.

Some modfications to tests where needed to account for the start of the
new sequence, and some more descriptive variable names used to help the
tests be more readable