$number_of_nodes formula variable
There were a number of concerns raised with how we were "exploding" each
app instance into it's own resource usage row; It was a point of
confusion in the view query and was artifically creating our dataset
much larger than it was.

The original idea for one-row-per-instance was to make the calculations
much easier to reason about, however we now have a significantly more
complex setup with variables in component formulas and it makes sense
that we treat the number of instances as a property for fomulas just
like any other... for example the redis service may make use of this
variable for calculating costs based on horizontal scaling.

Without the need to calculate instance rows in the resource_usage data
we take the opertunity to continue cleaning up:

* "guid" fields are now prefixed with what they are a guid of (ie
  resource_guid and event_guid)
* the hacky id concatination used to created a unique index is dropped
  in favour of using the event_guid (faster, cleaner)
* indexes are setup on all fields that are queried
* we intend on having resource "usage" and resource "billing" endpoints so we use the name
  resource_usage for the view to reflect it's use (props to henry there -
  he did the hard work)