Minimum charge of 1p per resource
Fractions of a penny cause misleadingly rounded prices and having high
precision prices complicates the rules around vat calculations
significantly so (at least for now) we are hard coding a minimum spend
of 1p per resource.

It's possible we may have to revisit this since this make extremly short
duration resources much more expensive. For example some kind of
auto-scaler that rapidly causes scale events several times a second
would get real expensive real fast... but it could be argued that such
rapid scheduling events would be putting a strain on the platform that
justifies the minimum spend and deters such actions.

In order to add a test for this to the api_test the /events endpoint was
switched to returning price in pence (it was the odd one out, all the
others already did). So all the test result values have been x100.