Use natural primary key for exchange rates
It was previously possible to multiple exchange rates with the same
valid_from, which would be an error.

We resolve this by replacing the table's artifical `id` key with the
natural composite primary key (code, valid_from) and add a test for the
new behaviour.

The tests were erroronously creating duplicates due to the
default/bootstrapped GBP currency ... to avoid this the bootstrapped
currency is now created at the epoch time.