Remove the temporary UI and browser oauth flow
The UI code had not kept pace with the rest of the codebase and it has
outlived it's usefulness.

The HTML rendered versions of the endpoints were useful to experiment
with the data and generate rough reports but we have no intension of
maintaining a user-facing UI for this application here (that's what
paas-admin is for), it is purely a JSON API so we remove all the HTML
rendering code.

Since we won't need the browser oauth flow (you can get a token directly
from uaa/cf) we can remove this code too.