Constraints, Performance and Declarative plan cfg
We want to be able to specify configuration of pricing plans, currency
and VAT on startup rather than via an API so that we can set it in our
manifest along with all the other configuration.

This is a big refactor that pulls together all the recent work to:

* Enable loading a JSON config file that specifies the plans by treating
all the pricing data tables as ephemeral and destroying them on each
refresh
* Improves query performance by ~100x by move as much as possible into
the hourly refresh/generation so that each query only does the work
required.
* Strips out all unneccasry api endpoints, we only need the
"billing_events" and "forecast_events" and possibly "usage_events"
* Improve the test readability so that it is clear what tests are doing
what and throw away all the confusing tests.