PoC scripts to help update pricing_plans
We created a little ruby script to help populate the pricing_plan data
for the RDS services by combining details from the paas-cf rds-broker
manifest, cf api and a ruby gem that fetches aws prices programatically.

It's far from perfect but still very useful as a proof of concept for
something more automated to either update or check prices against the
provider costs.