Improve the Collector/Fetcher interface
Our previous interfaces required that internals of our DB/Schema/Store
is exposed and extended for each new Fetcher, which was coupling the
code.

As a step towards breaking that coupling and having a simpler interface
for adding Fetchers this introduces an interface for Storing generic
events, and moves the responsibility of Storing the events to the
Collector from the Fetcher.