Replace gomock with counterfiter
We had a mix of gomock and counterfiter based mocks which was a bit
confusing, but more importantly gomock was problematic to install for
some reason.

This standandizes on the counterfiter mocking (we had more tests written
this way). The Makefile is used instead of go generate to ensure that
the mocks are regenerated when required.