AT LEAST ONE token key should verify the token
Each UAA client can have multiple token signing keys, and only ONE of
them needs to be able to verify the token signature. This is because as
token keys are rotated you may have have the token signed by a different
key.

This changes the behaviour from always using the first key to checking
that AT LEAST ONE key can verify the token.