make smoke
All the existing unit tests are mocked out so don't catch issues with
the real setup or even test if it can build at all. As a very rough
sanity test to see if everything is wired up correctly this adds a `make
smoke` task which can be run against a real environment.

We will extend this in the future with more comprehensive integration
tests, but for now it just checks that it can collect some app events
and hit the http endpoints without getting errors.

The smoke tests cannot run without a dev environment so are not included
in travis, they are only for manual use for now.