README: document configuration and usage
We update the README to reflect the recent changes to configuration and
API usage.

Adds documentation for all of the configuration options including
the new JSON config file for plans.

Documents the API endpoints and provides some curl examples to make
usage clearer.

Adds a pretty picture to help explain the architecture of the thing and
the relationship between the various eventio interfaces.