Implementing PricingPlanReader for EventStore
We want to be able to fetch the configured pricing plans for a given
date range from the public API to make working with the pricing
calculator easier.

This implements the GetPricingPlans() method of
eventio.PricingPlanReader for the EventStore to allow querying the
pricing_plans data currently configured.

It also add the PricingPlanHandler to the eventserver mounted at
"/pricing_plans".  The endpoint is exposed without authentication since
it will be used by the public (pricing calc) and does not expose
anything any more sensitive than is already visible in our public
manifests.