Use resource_type 'service' for all service events
We were erronously using the service_label as the resource_type in
events. The intention was that the resource_type would let you know what
kind of thing the resource_guid is referencing.

This updates the generation of the events table and tests to ensure that
all services use the 'service' type for resouce_type fields.