pg_size_bytes function
Postgres 9.6 adds a pg_size_bytes function for converting human readable
byte sizes into bytes integer values (ie 1KB => 1024).

We are unable to upgrade to 9.6 yet so we have created an implementation
of this function in plpgsql to use in the meantime. The tests have been
copied from upstream:

- https://github.com/postgres/postgres/blob/REL9_6_0/src/test/regress/expected/dbsize.out#L38-L146

This function can be removed once we have upgraded.