Inject compose scale events into service events
Compose deplpoyments can auto-scale which means the data we have at the
point of a CREATED service_usage_event can become stale.

To resolve this, we collect the "scaling" events from compose and inject
them into our normalized service events to enrich the events with the
data about the current memory/storage values.

We have to order by `created_at` in addition to `event_sequence` because
service_usage_event and compose_audit_event have different sequence IDs.

An additional sub-SELECT is used in order to lookup the correct
resource_name, memory_in_mb, and storage_in_mb values from the last
compose_audit_event or, if there aren't any, the original plan. This
ensures that we don't go back to using the plan if an UPDATED event
happens that doesn't result in the Compose service actually changing
size.