Add new encrypted plans for RDS
We want to offer only encrypted plans, as it has no cost and there
is no reason to use unencrypted.

We add all the corresponding missing plans that were not encrypted
in the PR https://github.com/alphagov/paas-cf/pull/1375:

 - small-5.7
 - medium-5.7
 - large-5.7
 - xlarge-5.7
 - tiny-9.5
 - small-9.5
 - medium-9.5
 - large-9.5
 - xlarge-9.5

The pricing is exactly the same than the unencrypted version.

The GUID can be retrieved doing:

    cf curl /v2/service_plans | jq -r '.resources[] | "\(.metadata.guid) \(.entity.name)"'