Handle Compose-Events newer than CF-Events
When calculating STARTED events for resources that we have never seen a
true STARTED/CREATED event for, the earlist date we can use is the date
that we started collecting the data.

We had events for compose that predate our collection of cf events.

This led to a situation where a compose event became the first seen
STARTED event for a service ... this was never expected to happen and so
compose events always fetch their name from the "previous" event in the
WINDOW (to handle renames gracefully).... as a result we had events
without a resource name, which is a constraint error and so caused the
application to shutdown as it believed it was operating with bad data
(which it was).

To resolve this, we fetch the name from the paired service_usage_event
in the case when the compose event is newer.

A test was added to catch regressions in this area.