Add `cfstore` package to billing app
This commit adds the new `cfstore` package to the billing app. The
purpose of this package is to enable the collection of service and
service plan data from Cloud Foundry and storing this data within the
billing database.

The consumers of this are the pricing calculator and the billing app
itself. Due to the names of the services and plans being used
as a method of identifing which plan to use, this was clearly a fragile
approach and when a plan was renamed the result would be the GUID being
displayed which has very little meaning to our users.