Allow token decryption
We've got a need to extract the token in more places now. It has been
previously done as part of the getVerifiedScopes() method, where it
didn't actually save the outcome for the future to come.

We've moved it outside of that method and saving the state in the
authorizer for other uses. If the method is called twice in the raw,
it'd reuse already prepared claims, in order to save some time and
resources.