Implement the Organisations functionality
We'd like to check if the user has either billing or org manager role in
requested organisation before they receive the information. In order to
achieve that, we're asking the CF API to provide us with the
organistations for that user where they have the desired role, and then
iterating over them to see if the requested guids are in fact listed.

We're removing the Spaces functionality, as in our belief, it is no
longer required nor used.