Log error on fail to find service for service_plan
If we cannot find a service for service_plan then we must skip it as we
will be unable to build the relationship between plan and service.

While it's possible that this occurs during a race between fetching
plans/services and deleting plans/services, hopefully in reality it is
rare and given this historic data "enhances" the event data by providing
names for guids, but is not technically "required" this should be
acceptable.