Log error if authentication failed
If there is any error during credential verification, we add
the root error cause to the raised error so we can troubleshoot
it as needed.

The error would be expose to the caller of the API.

We do not use a constant to keep the errors, but generate them
dinamically appending the error message. We do not use these constants
anywere, and building dynamically allows to append the actual error
to the propagated error message.