Verify token with the key that generated it
UAA might support multiple different signing keys for the token
verification in order to support signing key rotation[1]

The user token would contain a `kid` attribute to identify the
corresponding key that generated that token.

In our uaa_authorizer.go we were processing the keys, returning to
the `jwt.ParseWithClaims` the first one that matched the right
algorithm.

But that is not enough, because if the first key happens
to be a different one than the used to generate the token, the
verification would fail because it is not the right key.
We need to instead return the key that also has the same Key ID
(`kid`) than the token.

Sadly there was no testing around this functionality and integration
with UAA at all, so we had to implement some basic testing around
the composeClaims() methods. The example fixture RSA keys have been
extracted from some development environments, by querying UAA
`/token_keys` endpoint and directly picking the private key.

[1] http://docs.cloudfoundry.org/api/uaa/version/4.19.0/index.html#token-key-s

test