ensure deterministic ordering from getUsageEvents
We want to ensure that the results that come back from getUsageEvents is
in a known order, so that responses don't jump around and so that it is
easier to write tests for.

Currently the order is based on the timestamp only, but this is not good
enough when two events have the same timestamp.

We now add ordering by the event_guid too so we can rely on the order.