resolve compute/staging event clash
When we added the processing of STAGING_START / STAGING_STOP events we
attempted to roll them up into the regular app START / STOP events to
simplify the bills.

Unfortunatly this introduced a bug where a STAGING_STOP event could be
interperated as an application STOP event in certain circumstances (when
it came in late or at an identical time). This resulted in many
applications being undercharged.

There was also another error in the staging processing that used the
number_of_nodes from the target application, when in fact staging only
eveer uses a single node.

To resolve these, "staging" has been broken out into it's own "plan"
with it's own "plan_guid" in the same way that tasks are, and we then
partition by both the resource_guid AND the plan_guid when calculating
the ranges to ensure that only related events are paired up.

This may mean the bills are a bit more noisey as there will be more
events than previously (extra ones for "staging")