adds temporary event_type field during processing
I recently "fixed" an issue where START/STOP app events were getting mixed
up with STAGING_START/STAGING_STOP events by including the plan_guid in
the partition query so that START/STOP events were only considered from
the same plan.

This did not take into account that resources can change their plan_guid
during UPDATE events, and we did not have a test to catch this.

A previous commit added a test to illutrate this problem and this commit
introduces an "event_type" field during the processing of events that
can be used to isolate exactly which event types should be getting
paired up to resolve the issue.