Add org names to billing data
We want to be able to show org names, not GUIDs, in situations where an
org may no longer exist.

To do this we firstly have to collect the org names from the CF API and
persist them in the billing database. These names are valid from
whenever the org was last updated. The valid names are then injected
into the billable events.

We also had to update the forecasting functionality as they use the same
data structures.

Note: this currently has no user-visible impact, because no endpoint in
our admin app currently shows deleted orgs.