Allow targeted package tests to be run from Make
This commit allows PACKAGE and ACTION envvars (or Make-vars) to focus
the Makefile-driven tests.

PACKAGE is injected as the final param to `ginkgo`, allowing one to run:

    make PACKAGE=cfstore test

ACTION is injected as the verb that `ginkgo` will run. e.g.

    make ACTION=watch test

These are implemented here as the Makefile sets up various testing
envvars which are slightly annoying to replicate at the CLI - making the
Makefile the better entrypoint for this testing.