Add space names to billing data
We want to be able to show space names, not GUIDs, in situations where
a space may no longer exist.

To do this we firstly have to collect the space names from the CF API
and persist them in the billing database. These names are valid from
whenever the space was last updated. The valid names are then injected
into the billable events.

We also had to update the forecasting functionality as they use the same
data structures.