Update service plan GUIDs in config
We needed to update the the config with the service plan GUIDs that are
provided by the broker service plan rather than the ones generated by
Cloud Foundry.

This are unique IDs that are consistent across all the deployments,
so all the formulas in our service plans would work everywhere.

In following commits we update the billing platform to use this
unique broker ID, that is stored in the `service_plans` table.