Match the pricing_plan by service_plan.unique_id
We want to be able to match a event with the corresponding
pricing plan based on the unique_id stored in the service_plan
table.

For that, we enrich the table event with a field `unique_id`, which
will be populated with the value of the `service_plan.unique_id`
when the service type is `service`.

For other event types, `event.unique_id` would get the same value
than plan_id, so that existing plans and the app/staging plans
would continue working.

Adding the events.unique_id also makes easier to check if all
the plans exist in `checkPlanConsistency()`, as we only need
to check that a plan exists for that field.

This new logic might break the computation of the events if there
is an event referring to a `plan_guid` that does not exist in
the `service_plans` table. This is not a problem, because the
event computation is disconnected from the billing generation,
and this table should be periodically refreshed every 20 minutes.

Note: We had to add a casting of service_plan.unique_id::uuid. We
don't know the reason for this, as the table field
is declared as uuid.