Forecast events include the plan_unique_id
Now we are using the pricing plan unique_id to identify the plans.
In the case of the forecast events, we will also pass the
same ID for both the `plan_guid` and the `plan_unique_id`.

The consumer of the forecast events (paas-admin pricing-calculator)
would pass the IDs previously reported by the `/pricing_plans`
endpoint.

Note that after this change, any old URL with the old IDs
would stop working.